# graphql-typescript-definitions performance improvements