## Description
Prevent `graphql-typescript-definitions` queuing up a bunch of mostly redundant work. tl;dr is that local/Spin workspaces may take 2+ minutes of 100% CPU to complete type def work.

Examples of the churn generated are provided below in collapsed `details` blocks.

### What impact does this have?
I ran through before/after for a couple of scenarios in Shopify's largest app. All scenarios start with sewing-kit running a development server using:

```sh
yarn dev --focus Home
```

#### Schema refresh (`yarn refresh-graphql --mode production --no-definitions`)

Without these improvements:
* `Unable to find any GraphQL type definitions for` is output 90 times
* `Built GraphQL type definitions` is output 20 times
* JS compilation is initiated 11 times, including one 26 second client compile

With these improvements:
* `Unable to find any GraphQL type definitions for` no longer appears
* `Built GraphQL type definitions` appears once
* Still a couple of `no such file or directory` errors (fixed by https://github.com/Shopify/sewing-kit/pull/2846)
* 4 compilations initiated (2 seconds max compile time)

<details>
<summary>Dev server output before this PR</summary>

```
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/hsCodesService.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/mailbox.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/shopifyChat.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/appStore.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/bourgeois.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/banking.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/argus.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/merchantRisk.json
[client] Compiling
[server] Compiling
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open './build/graphql/schemas/core-schema-unions-and-interfaces.json'

[server] Build failed.
[server] Compiling
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open './build/graphql/schemas/core-schema-unions-and-interfaces.json'

[client] Build failed.
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open './build/graphql/schemas/core-schema-unions-and-interfaces.json'

[server] Build failed.
[client] Compiling
[graphql] Built GraphQL schema types
[error]
      Unable to find any GraphQL type definitions for the following pointers:
          - build/graphql/schemas/core.json
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open './build/graphql/schemas/core-schema-unions-and-interfaces.json'

[client] Build failed.
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[server] Compiling
[client] Compiling
[server] Compiled with latest changes (2s)
[graphql] Built GraphQL schema types
[server] Compiling
[graphql] Built GraphQL type definitions
[client] Compiled with latest changes (16s)
[client] Creating consolidated manifest
[client] Completed manifest creation
[client] Compiling
[server] Running with latest changes
[server] Compiled with latest changes (9s)
[init] listening on 192.168.64.1:40020
[server] Compiling
[graphql] Built GraphQL type definitions
[server] Compiled with latest changes (4s)
[client] Compiled with latest changes (26s)
[client] Creating consolidated manifest
[client] Completed manifest creation
[server] Compiling
[server] Running with latest changes
[client] Compiling
[init] listening on 192.168.64.1:40020
[server] Compiled with latest changes (6s)
[client] Compiled with latest changes (2s)
[client] Creating consolidated manifest
[client] Completed manifest creation
[server] Running with latest changes
[init] listening on 192.168.64.1:40020
```
</details>

<details>
<summary>Dev server output after this PR</summary>

```
[server] Compiling
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open '/Users/gord/src/github.com/Shopify/web/build/graphql/schemas/core-schema-unions-and-interfaces.json'

[server] Build failed.
[client] Compiling
./build/graphql/schemas/core-schema-unions-and-interfaces.json
Module build failed: Error: ENOENT: no such file or directory, open '/Users/gord/src/github.com/Shopify/web/build/graphql/schemas/core-schema-unions-and-interfaces.json'

[client] Build failed.
[graphql] Built GraphQL type definitions
[server] Compiling
[client] Compiling
[server] Compiled with latest changes (2s)
[client] Compiled with latest changes (2s)
[client] Creating consolidated manifest
[client] Completed manifest creation
[server] Running with latest changes
[graphql] Built GraphQL type definitions
[init] listening on 192.168.64.1:40020
```
</details>

#### Fast-forward from stale head
To simulate a pull on a stale branch, I did this:
```
git reset --hard HEAD~20
git pull
```

Without these improvements:
* `Built GraphQL type definitions` is output 31 times
* Almost 6 minutes is spent compiling client JS (compilation is _very_ slow because so much GQL type generation is silently churning in the background)

After these improvements:
* `Built GraphQL type definitions` appears twice
* More compilations are triggered! (this is expected because they last ~4 seconds, so Chokidar has more time to pick up changes as Git churns through commits)
* A few JS compilation errors are thrown (expected because files appear/disappear during pulls 🤷 ) 

<details>
<summary>Dev server output before these improvements</summary>

```
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[server] Compiling
[client] Compiling
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[graphql] Built GraphQL type definitions
[server] Compiled with latest changes (2m52s)
[server] Compiling
[server] Compiled with latest changes (1s)
[client] Compiled with latest changes (2m54s)
[client] Creating consolidated manifest
[client] Completed manifest creation
[client] Compiling
[server] Running with latest changes
[client] Compiled with latest changes (2s)
[client] Creating consolidated manifest
[client] Completed manifest creation
…